backend/app/requirements.txt
alembic
asyncpg
httpx-oauth
fastapi
fastapi-users[sqlalchemy,oauth]
sqlalchemy-utils
psycopg2-binary
python-multipart
uvicorn